backend/requirement.txt
django==2.1.3
django-templated-mail==1.1.1
PyJWT==1.7.0
djangorestframework-jwt==1.11.0
psycopg2-binary==2.7.7
djangorestframework==3.9.0
markdown==3.0.1
django-filter==2.0.0
django-cors-headers==2.4.0
Pillow
fcm-django
django-shopify-webhook
enum-compat==0.0.2
eventlet==0.21.0
greenlet==0.4.12
python-engineio
python-socketio
pytz==2018.7
six==1.10.0
django-dotenv
boto3
authy
